+++ src/electron/output/Output.ts
@@ -1,0 +1,6 @@
+import { BrowserWindow } from "electron"
+
+export class Output {
+    static window: BrowserWindow
+    static previewWindow: BrowserWindow
+}
