+++ src/electron/output/OutputHelper.ts
@@ -1,15 +1,15 @@
 import { BrowserWindow } from "electron"
-import { OutputBounds } from "./OutputBounds"
-import { OutputIdentify } from "./OutputIdentify"
-import { OutputSend } from "./OutputSend"
-import { OutputValues } from "./OutputValues"
-import { OutputVisibility } from "./OutputVisibility"
-import { OutputLifecycle } from "./OutputLifecycle"
-import { Message } from "../../../types/Socket"
-import { toApp } from "../.."
-import { CaptureTransmitter } from "../CaptureTransmitter"
-import { updatePreviewResolution } from "../capture"
-import { OUTPUT } from "../../../types/Channels"
+import { OutputBounds } from "./helpers/OutputBounds"
+import { OutputIdentify } from "./helpers/OutputIdentify"
+import { OutputSend } from "./helpers/OutputSend"
+import { OutputValues } from "./helpers/OutputValues"
+import { OutputVisibility } from "./helpers/OutputVisibility"
+import { OutputLifecycle } from "./helpers/OutputLifecycle"
+import { Message } from "../../types/Socket"
+import { toApp } from ".."
+import { CaptureTransmitter } from "./CaptureTransmitter"
+import { updatePreviewResolution } from "./capture"
+import { OUTPUT } from "../../types/Channels"
 
 export class OutputHelper {
     static receiveOutput(_e: any, msg: Message) {
